Refactor code to use buttons for user input
The code was adjusted to declare a winner for the first person to win 5 rounds instead of the best out of 5. Buttons for user input are disabled once a game winner has been declared to effectively end the game.

Buttons were added to the HTML file to serve as user input methods as opposed to requesting text with a prompt.